+++ src/components/Button/Button.scss
@@ -3,7 +3,6 @@
 .button {
   background-color: $main-red;
   border-radius: 8px;
-
   span {
     color: white;
     font-weight: 600;
