+++ src/pages/Application/ApplicationManagement/ApplicationManagement.scss
@@ -1,4 +1,4 @@
-@import "../../styles/common";
+@import "../../../styles/common.scss";
 
 .application-management {
   width: 75%;
