+++ src/pages/Application/ApplicationPage/Question/Question.scss
