+++ src/pages/Application/ApplicationPage/Questions/Questions.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/common";
+@import "../../../../styles/common.scss";
 
 .application-questions {
   .application-item {
